reorganized files, added empty chat message
