+++ SH4RE/SH4RE/Messages/Components/MessageView.swift
@@ -1,0 +1,42 @@
+//
+//  MessageView.swift
+//  SH4RE
+//
+//  Created by Hannah Brooks on 2023-02-09.
+//
+
+import SwiftUI
+
+struct MessageView: View {
+    let message: ChatMessage
+    
+    var body: some View {
+        VStack {
+            if message.fromId == getCurrentUserUid() {
+                HStack {
+                    Spacer()
+                    HStack {
+                        Text(message.text)
+                            .foregroundColor(.white)
+                    }
+                    .padding()
+                    .background(Color.primaryDark)
+                    .cornerRadius(8)
+                }
+            } else {
+                HStack {
+                    HStack {
+                        Text(message.text)
+                            .foregroundColor(.black)
+                    }
+                    .padding()
+                    .background(.white)
+                    .cornerRadius(8)
+                    Spacer()
+                }
+            }
+        }
+        .padding(.horizontal)
+        .padding(.top, 8)
+    }
+}
