+++ SH4RE/SH4RE/Messages/MessagesChat.swift
@@ -6,155 +6,6 @@
 //
 
 import SwiftUI
-import Firebase
-
-class ChatLogViewModel: ObservableObject {
-    
-    @Published var chatText = ""
-    @Published var errorMessage = ""
-    
-    @Published var chatMessages = [ChatMessage]()
-    
-    var chatUser: ChatUser?
-    
-    init(chatUser: ChatUser?) {
-        self.chatUser = chatUser
-        
-        fetchMessages()
-    }
-    
-    var firestoreListener: ListenerRegistration?
-    
-    func fetchMessages() {
-        guard let fromId = FirebaseManager.shared.auth.currentUser?.uid else { return }
-        guard let toId = chatUser?.uid else { return }
-        firestoreListener?.remove()
-        chatMessages.removeAll()
-        firestoreListener = FirebaseManager.shared.firestore
-            .collection(FirebaseConstants.messages)
-            .document(fromId)
-            .collection(toId)
-            .order(by: FirebaseConstants.timestamp)
-            .addSnapshotListener { querySnapshot, error in
-                if let error = error {
-                    self.errorMessage = "Failed to listen for messages: \(error)"
-                    print(error)
-                    return
-                }
-
-                querySnapshot?.documentChanges.forEach({ change in
-                    if change.type == .added {
-                        let data = change.document.data()
-                        let fromId = data["fromId"] as? String ?? ""
-                        let toId = data["toId"] as? String ?? ""
-                        let text = data["text"] as? String ?? ""
-                        let date = data["timestamp"] as? Date ?? Date()
-                        let cm = ChatMessage(id: change.document.documentID, fromId: fromId, toId: toId, text: text, timestamp: date)
-                        self.chatMessages.append(cm)
-                    }
-                })
-
-                DispatchQueue.main.async {
-                    self.count += 1
-                }
-            }
-    }
-    
-    func handleSend() {
-        guard let fromId = FirebaseManager.shared.auth.currentUser?.uid else { return }
-
-        guard let toId = chatUser?.uid else { return }
-
-        let document = FirebaseManager.shared.firestore.collection(FirebaseConstants.messages)
-            .document(fromId)
-            .collection(toId)
-            .document()
-
-        let msg = ChatMessage(id: nil, fromId: fromId, toId: toId, text: chatText, timestamp: Date())
-
-        try? document.setData(from: msg) { error in
-            if let error = error {
-                print(error)
-                self.errorMessage = "Failed to save message into Firestore: \(error)"
-                return
-            }
-
-            self.persistRecentMessage()
-
-            self.chatText = ""
-            self.count += 1
-        }
-
-        let recipientMessageDocument = FirebaseManager.shared.firestore.collection("messages")
-            .document(toId)
-            .collection(fromId)
-            .document()
-
-        try? recipientMessageDocument.setData(from: msg) { error in
-            if let error = error {
-                print(error)
-                self.errorMessage = "Failed to save message into Firestore: \(error)"
-                return
-            }
-        }
-    }
-    
-    private func persistRecentMessage() {
-        guard let chatUser = chatUser else { return }
-
-        guard let uid = FirebaseManager.shared.auth.currentUser?.uid else { return }
-        guard let toId = self.chatUser?.uid else { return }
-
-        let document = FirebaseManager.shared.firestore
-            .collection(FirebaseConstants.recentMessages)
-            .document(uid)
-            .collection(FirebaseConstants.messages)
-            .document(toId)
-
-        let data = [
-            FirebaseConstants.timestamp: Date(),
-            FirebaseConstants.text: self.chatText,
-            FirebaseConstants.fromId: uid,
-            FirebaseConstants.toId: toId,
-            "name": chatUser.name
-        ] as [String : Any]
-
-        document.setData(data) { error in
-            if let error = error {
-                self.errorMessage = "Failed to save recent message: \(error)"
-                print("Failed to save recent message: \(error)")
-                return
-            }
-        }
-        guard let currentUser = FirebaseManager.shared.currentUser else {
-            return
-        }
-        let recipientRecentMessageDictionary = [
-            FirebaseConstants.timestamp: Date(),
-            FirebaseConstants.text: self.chatText,
-            FirebaseConstants.fromId: uid,
-            FirebaseConstants.toId: toId,
-            "name": currentUser.name
-        ] as [String : Any]
-        
-        let recipDoc = FirebaseManager.shared.firestore
-            .collection(FirebaseConstants.recentMessages)
-            .document(toId)
-            .collection(FirebaseConstants.messages)
-            .document(uid)
-
-        recipDoc.setData(recipientRecentMessageDictionary) { error in
-            if let error = error {
-                self.errorMessage = "Failed to save recent message: \(error)"
-                print("Failed to save recent message: \(error)")
-                return
-            }
-        }
-    }
-    
-    @Published var count = 0
-}
-
 
 struct MessagesChat: View {
     
@@ -227,56 +78,6 @@ struct MessagesChat: View {
         }
         .padding(.horizontal)
         .padding(.vertical, 8)
-    }
-}
-    
-
-
-struct MessageView: View {
-    let message: ChatMessage
-    
-    var body: some View {
-        VStack {
-            if message.fromId == getCurrentUserUid() {
-                HStack {
-                    Spacer()
-                    HStack {
-                        Text(message.text)
-                            .foregroundColor(.white)
-                    }
-                    .padding()
-                    .background(Color.primaryDark)
-                    .cornerRadius(8)
-                }
-            } else {
-                HStack {
-                    HStack {
-                        Text(message.text)
-                            .foregroundColor(.black)
-                    }
-                    .padding()
-                    .background(.white)
-                    .cornerRadius(8)
-                    Spacer()
-                }
-            }
-        }
-        .padding(.horizontal)
-        .padding(.top, 8)
-    }
-}
-
-private struct Placeholder: View {
-    var body: some View {
-        HStack {
-            Text("Message")
-                .keyboardType(.numberPad)
-                .foregroundColor(.darkGrey)
-                .font(.body)
-                .padding(.leading, 5)
-                .padding(.top, -4)
-            Spacer()
-        }
     }
 }
